replacing readonlystate with state and the original state to MutableState to match the coordinates nameing
