--- core/src/main/java/com/novoda/landingstrip/ReadOnlyState.java
@@ -1,10 +1,0 @@
-package com.novoda.landingstrip;
-
-interface ReadOnlyState {
-
-    int position();
-
-    float offset();
-
-    int fastForwardPosition();
-}
